Created C64 port
